+++ contracts/libraries/Events.sol
@@ -4,12 +4,12 @@ pragma solidity 0.8.28;
 import "../LaunchPoolFactoryUpgradeable.sol";
 
 library Events {
-    event NewProject(uint256 indexed projectId, address indexed rewardToken, address indexed owner, LaunchPoolFactoryUpgradeable.PoolMetadata metadata);
-    event NewLaunchPool(uint256 indexed projectId, address indexed launchPool, uint256 version);
-    event ProjectStatusUpdated(uint256 indexed projectId, LaunchPoolFactoryUpgradeable.ProjectStatus status);
-    event PoolMetadataUpdated(uint256 indexed projectId, LaunchPoolFactoryUpgradeable.PoolMetadata metadata);
-    event PoolFunded(uint256 indexed projectId, address indexed pool);
-    event ProjectOwnershipTransferred(uint256 indexed projectId, address indexed previousOwner, address indexed newOwner);
+    event NewProject(uint32 indexed projectId, address indexed rewardToken, address indexed owner, LaunchPoolFactoryUpgradeable.PoolMetadata metadata);
+    event NewLaunchPool(uint32 indexed projectId, address indexed launchPool, uint256 version);
+    event ProjectStatusUpdated(uint32 indexed projectId, LaunchPoolFactoryUpgradeable.ProjectStatus status);
+    event PoolMetadataUpdated(uint32 indexed projectId, LaunchPoolFactoryUpgradeable.PoolMetadata metadata);
+    event PoolFunded(uint32 indexed projectId, address indexed pool);
+    event ProjectOwnershipTransferred(uint32 indexed projectId, address indexed previousOwner, address indexed newOwner);
     event FactoryUpgraded(address indexed implementation);
     event RemainingRewardsWithdrawn(address indexed owner, uint256 amount);
 }
